# kanban/kanban.business/kanban.business.impl/src/main/java/com/kanban/business/common/GenericResponse.java
package com.kanban.business.common;

public class GenericResponse {

	private String message;
	private Object data;

	public GenericResponse() {
		message = "ok";
	}

	public String getMessage() {
		return message;
	}

	public void setMessage(String message) {
		this.message = message;
	}

	/**
	 * @return the data
	 */
	public Object getData() {
		return data;
	}

	/**
	 * @param data the data to set
	 */
	public void setData(Object data) {
		this.data = data;
	}

}