# kanban/kanban.business/kanban.business.impl/src/main/java/com/kanban/business/dto/LoginDto.java
/**
 * 
 */
package com.kanban.business.dto;

/**
 * @author jhernandez
 *
 */
public class LoginDto {

	private String identification;
	private String password;

	/**
	 * @return the identification
	 */
	public String getIdentification() {
		return identification;
	}

	/**
	 * @param identification the identification to set
	 */
	public void setIdentification(String identification) {
		this.identification = identification;
	}

	/**
	 * @return the password
	 */
	public String getPassword() {
		return password;
	}

	/**
	 * @param password the password to set
	 */
	public void setPassword(String password) {
		this.password = password;
	}

}